+++ src/main/java/se/subject/repositories/IUserRepository.java
@@ -24,6 +24,6 @@ import org.springframework.stereotype.Repository;
 import se.subject.entities.User;
 
 @Repository
-public interface IUserRepository extends CrudRepository<User, Long> {
+public interface IUserRepository extends CrudRepository<User, Integer> {
 	User findByEmail(String email);
 }
